requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==2.1.15
django-pyodbc-azure==2.1.0.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.0
pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
